requirements.txt
Django==1.8
python-twitter==3.3
mysqlclient==1.3.6
djangorestframework==3.1.3
unicodecsv==0.14